+++ R/drdid_panel.R
@@ -10,7 +10,7 @@ NULL
 #' @param y1 An \eqn{n} x \eqn{1} vector of outcomes from the post-treatment period.
 #' @param y0 An \eqn{n} x \eqn{1} vector of outcomes from the pre-treatment period.
 #' @param D An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).
-#' @param covariates An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score and regression estimation.
+#' @param covariates An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score and regression estimation. Please add a vector of constants if you want to include an intercept in the models.
 #' If covariates = NULL, this leads to an unconditional DiD estimator.
 #' @param i.weights An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights. The weights are normalized and therefore enforced to have mean 1 across all observations.
 #' @param boot Logical argument to whether bootstrap should be used for inference. Default is FALSE.
@@ -55,7 +55,7 @@ NULL
 #' unit_random <- sample(1:nrow(eval_lalonde_cps), 5000)
 #' eval_lalonde_cps <- eval_lalonde_cps[unit_random,]
 #' # Select some covariates
-#' covX = as.matrix(cbind(eval_lalonde_cps$age, eval_lalonde_cps$educ,
+#' covX = as.matrix(cbind(1, eval_lalonde_cps$age, eval_lalonde_cps$educ,
 #'                        eval_lalonde_cps$black, eval_lalonde_cps$married,
 #'                        eval_lalonde_cps$nodegree, eval_lalonde_cps$hisp,
 #'                        eval_lalonde_cps$re74))
@@ -77,15 +77,21 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   n <- length(D)
   # generate deltaY
   deltaY <- as.vector(y1 - y0)
-  # Add constant to covariate vector
-  int.cov <- as.matrix(rep(1,n))
-  if (!is.null(covariates)){
-    if(all(as.matrix(covariates)[,1]==rep(1,n))){
-      int.cov <- as.matrix(covariates)
-    } else {
-      int.cov <- as.matrix(cbind(1, covariates))
-    }
+  # Covariate vector
+  if(is.null(covariates)){
+    int.cov <- as.matrix(rep(1,n))
+  } else{
+    int.cov <- as.matrix(covariates)
   }
+
+  # int.cov <- as.matrix(rep(1,n))
+  # if (!is.null(covariates)){
+  #   if(all(as.matrix(covariates)[,1]==rep(1,n))){
+  #     int.cov <- as.matrix(covariates)
+  #   } else {
+  #     int.cov <- as.matrix(cbind(1, covariates))
+  #   }
+  # }
 
   # Weights
   if(is.null(i.weights)) {
@@ -96,27 +102,32 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   i.weights <- i.weights/mean(i.weights)
   #-----------------------------------------------------------------------------
   #Compute the Pscore by MLE
-  #pscore.tr <- stats::glm(D ~ -1 + int.cov, family = "binomial", weights = i.weights)
-  pscore.tr <- suppressWarnings(parglm::parglm(D ~ -1 + int.cov, family = "binomial", weights = i.weights))
+  pscore.tr <- suppressWarnings(fastglm::fastglm(
+                                x = int.cov,
+                                y = D,
+                                family = stats::binomial(),
+                                weights = i.weights,
+                                intercept = FALSE,
+                                method = 3
+  ))
+  class(pscore.tr) <- "glm" #this allow us to use vcov
   if(pscore.tr$converged == FALSE){
     warning("Propernsity score estimation did not converge.")
   }
   if(anyNA(pscore.tr$coefficients)){
     stop("Propensity score model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is a likely reason.")
   }
-  ps.fit <- as.vector(pscore.tr$fitted.values)
+  ps.fit <- fitted(pscore.tr) #as.vector(pscore.tr$fitted.values)
   # Avoid divide by zero
   ps.fit <- pmin(ps.fit, 1 - 1e-6)
-  #Compute the Outcome regression for the control group using wols
-  # reg.coeff <- stats::coef(stats::lm(deltaY ~ -1 + int.cov,
-  #                                    subset = D==0,
-  #                                    weights = i.weights))
+  W <- ps.fit * (1 - ps.fit) * i.weights
+  # Compute the Outcome regression for the control group using wols
   control_filter <- (D == 0)
   reg.coeff <- stats::coef(fastglm::fastglm(
-                            x = int.cov[control_filter, , drop = FALSE],
-                            y = deltaY[control_filter],
-                            weights = i.weights[control_filter],
-                            family = gaussian(link = "identity")
+    x = int.cov[control_filter, , drop = FALSE],
+    y = deltaY[control_filter],
+    weights = i.weights[control_filter],
+    family =  stats::gaussian(link = "identity")
   ))
   if(anyNA(reg.coeff)){
     stop("Outcome regression model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is a likely reason.")
@@ -142,6 +153,7 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   weights.ols <- i.weights * (1 - D)
   wols.x <- weights.ols * int.cov
   wols.eX <- weights.ols * (deltaY - out.delta) * int.cov
+  #XpX <- opt_crossprod(wols.x, int.cov, n)
   XpX <- base::crossprod(wols.x, int.cov)/n
   # Check if XpX is invertible
   if ( base::rcond(XpX) < .Machine$double.eps) {
@@ -152,8 +164,10 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
 
   # Asymptotic linear representation of logit's beta's
   score.ps <- i.weights * (D - ps.fit) * int.cov
-  Hessian.ps <- stats::vcov(pscore.tr) * n
+  #Hessian.ps <- solve(t(int.cov) %*% (W * int.cov)) * n
+  Hessian.ps <- chol2inv(chol(t(int.cov) %*% (W * int.cov))) * n
   asy.lin.rep.ps <-  score.ps %*% Hessian.ps
+
 
   # Now, the influence function of the "treat" component
   # Leading term of the influence function: no estimation effect
@@ -178,8 +192,19 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
   inf.cont.2 <- asy.lin.rep.ps %*% M2
   # Estimation Effect from beta hat (weighted OLS)
   M3 <-  base::colMeans(w.cont * int.cov)
-  # Now the influence function related to estimation effect of regressions
   inf.cont.3 <- asy.lin.rep.wols %*% M3
+
+  # # Batch multiple matrix multiplications for inf functions
+  # batch_results <- batch_matrix_operations(wols.eX, XpX.inv, score.ps, Hessian.ps, M1, M2, M3)
+  # # Now get the influence function related to the estimation effect related to beta's
+  # inf.treat.2 <- batch_results$inf_treat_2
+  # # Now the influence function related to estimation effect of pscores
+  # inf.cont.2 <- batch_results$inf_cont_2
+  # # Now the influence function related to estimation effect of regressions
+  # inf.cont.3 <- batch_results$inf_cont_3
+
+  # Influence function for the treated component
+  inf.treat <- (inf.treat.1 - inf.treat.2) / mean(w.treat)
 
   # Influence function for the control component
   inf.control <- (inf.cont.1 + inf.cont.2 - inf.cont.3) / mean(w.cont)
@@ -205,11 +230,11 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
       dr.boot <- mboot.did(dr.att.inf.func, nboot)
       # get bootstrap std errors based on IQR
       se.dr.att <- stats::IQR(dr.boot) / (stats::qnorm(0.75) - stats::qnorm(0.25))
-      # get symmtric critival values
+      # get symmetric critical values
       cv <- stats::quantile(abs(dr.boot/se.dr.att), probs = 0.95)
-      # Estimate of upper boudary of 95% CI
+      # Estimate of upper boundary of 95% CI
       uci <- dr.att + cv * se.dr.att
-      # Estimate of lower doundary of 95% CI
+      # Estimate of lower boundary of 95% CI
       lci <- dr.att - cv * se.dr.att
     } else {
       # do weighted bootstrap
@@ -217,11 +242,11 @@ drdid_panel <-function(y1, y0, D, covariates, i.weights = NULL,
                                n = n, deltaY = deltaY, D = D, int.cov = int.cov, i.weights = i.weights))
       # get bootstrap std errors based on IQR
       se.dr.att <- stats::IQR((dr.boot - dr.att)) / (stats::qnorm(0.75) - stats::qnorm(0.25))
-      # get symmtric critival values
+      # get symmetric critical values
       cv <- stats::quantile(abs((dr.boot - dr.att)/se.dr.att), probs = 0.95)
-      # Estimate of upper boudary of 95% CI
+      # Estimate of upper boundary of 95% CI
       uci <- dr.att + cv * se.dr.att
-      # Estimate of lower doundary of 95% CI
+      # Estimate of lower boundary of 95% CI
       lci <- dr.att - cv * se.dr.att
 
     }