+++ R/std_ipw_did_rc.R
@@ -12,7 +12,7 @@ NULL
 #' @param post An \eqn{n} x \eqn{1} vector of Post-Treatment dummies (post = 1 if observation belongs to post-treatment period,
 #'             and post = 0 if observation belongs to pre-treatment period.)
 #' @param D An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).
-#' @param covariates An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score estimation.
+#' @param covariates An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score estimation. Please add a column of ones if you want to include an intercept.
 #' If covariates = NULL, this leads to an unconditional DID estimator.
 #' @param i.weights An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights. The weights are normalized and therefore enforced to have mean 1 across all observations.
 #' @param boot Logical argument to whether bootstrap should be used for inference. Default is FALSE.
@@ -44,7 +44,7 @@ NULL
 #'
 #' @examples
 #' # use the simulated data provided in the package
-#' covX = as.matrix(sim_rc[,5:8])
+#' covX = as.matrix(cbind(1, sim_rc[,5:8]))
 #' # Implement normalized IPW DID estimator
 #' std_ipw_did_rc(y = sim_rc$y, post = sim_rc$post, D = sim_rc$d,
 #'                covariates= covX)
@@ -63,15 +63,13 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
   y <- as.vector(y)
   # post as vector
   post <- as.vector(post)
-  # Add constant to covariate vector
-  int.cov <- as.matrix(rep(1,n))
-  if (!is.null(covariates)){
-    if(all(as.matrix(covariates)[,1]==rep(1,n))){
-      int.cov <- as.matrix(covariates)
-    } else {
-      int.cov <- as.matrix(cbind(1, covariates))
-    }
+  # Covariate vector
+  if(is.null(covariates)){
+    int.cov <- as.matrix(rep(1,n))
+  } else{
+    int.cov <- as.matrix(covariates)
   }
+
 
   # Weights
   if(is.null(i.weights)) {
@@ -80,12 +78,19 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
   # Normalize weights
   i.weights <- i.weights/mean(i.weights)
   #-----------------------------------------------------------------------------
-  #Pscore estimation (logit) and also its fitted values
-  #PS <- suppressWarnings(stats::glm(D ~ -1 + int.cov, family = "binomial", weights = i.weights))
-  PS <- suppressWarnings(parglm::parglm(D ~ -1 + int.cov, family = "binomial", weights = i.weights))
-  ps.fit <- as.vector(PS$fitted.values)
+  # Pscore estimation (logit) and also its fitted values
+  PS <- suppressWarnings(fastglm::fastglm(
+                                  x = int.cov,
+                                  y = D,
+                                  family = stats::binomial(),
+                                  weights = i.weights,
+                                  intercept = FALSE,
+                                  method = 3
+  ))
+  ps.fit <- fitted(PS)
   # Do not divide by zero
   ps.fit <- pmin(ps.fit, 1 - 1e-6)
+  W <- ps.fit * (1 - ps.fit) * i.weights
   #-----------------------------------------------------------------------------
   #Compute IPW estimator
   # First, the weights
@@ -113,7 +118,8 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
   #-----------------------------------------------------------------------------
   # Asymptotic linear representation of logit's beta's
   score.ps <- i.weights * (D - ps.fit) * int.cov
-  Hessian.ps <- stats::vcov(PS) * n
+  #Hessian.ps <- stats::vcov(PS) * n
+  Hessian.ps <- chol2inv(chol(t(int.cov) %*% (W * int.cov))) * n
   asy.lin.rep.ps <-  score.ps %*% Hessian.ps
   #-----------------------------------------------------------------------------
   # Now, the influence function of the "treat" component
@@ -144,9 +150,9 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
   if (boot == FALSE) {
     # Estimate of standard error
     se.att <- stats::sd(att.inf.func)/sqrt(n)
-    # Estimate of upper boudary of 95% CI
+    # Estimate of upper boundary of 95% CI
     uci <- ipw.att + 1.96 * se.att
-    # Estimate of lower doundary of 95% CI
+    # Estimate of lower boundary of 95% CI
     lci <- ipw.att - 1.96 * se.att
 
     #Create this null vector so we can export the bootstrap draws too.
@@ -159,11 +165,11 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
       ipw.boot <- mboot.did(att.inf.func, nboot)
       # get bootstrap std errors based on IQR
       se.att <- stats::IQR(ipw.boot) / (stats::qnorm(0.75) - stats::qnorm(0.25))
-      # get symmtric critival values
+      # get symmetric critical values
       cv <- stats::quantile(abs(ipw.boot/se.att), probs = 0.95)
-      # Estimate of upper boudary of 95% CI
+      # Estimate of upper boundary of 95% CI
       uci <- ipw.att + cv * se.att
-      # Estimate of lower doundary of 95% CI
+      # Estimate of lower boundary of 95% CI
       lci <- ipw.att - cv * se.att
     } else {
       # do weighted bootstrap
@@ -171,11 +177,11 @@ std_ipw_did_rc <-function(y, post, D, covariates, i.weights = NULL,
                                 n = n, y = y, post = post, D = D, int.cov = int.cov, i.weights = i.weights))
       # get bootstrap std errors based on IQR
       se.att <- stats::IQR(ipw.boot - ipw.att) / (stats::qnorm(0.75) - stats::qnorm(0.25))
-      # get symmtric critival values
+      # get symmetric critical values
       cv <- stats::quantile(abs((ipw.boot - ipw.att)/se.att), probs = 0.95)
-      # Estimate of upper boudary of 95% CI
+      # Estimate of upper boundary of 95% CI
       uci <- ipw.att + cv * se.att
-      # Estimate of lower doundary of 95% CI
+      # Estimate of lower boundary of 95% CI
       lci <- ipw.att - cv * se.att
 
     }